Dino4Cells/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
backcall=0.2.0=py_0
bleach=4.0.0=pypi_0
bottleneck=1.3.2=pypi_0
ca-certificates=2020.10.14=0
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.6=pypi_0
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=py_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
faiss-gpu=1.7.1=py3.6_h080d439_1_cuda10.2
harmony-pytorch=0.1.6=pypi_0
htmlmin=0.1.12=pypi_0
idna=2.10=pypi_0
imagehash=4.2.1=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=4.6.3=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.6.3=pypi_0
jedi=0.17.2=py36_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jsonschema=3.2.0=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
kornia=0.5.10=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libfaiss=1.7.1=h4fe19ad_1_cuda10.2
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libopenblas=0.3.13=h4367d64_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
llvmlite=0.36.0=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
missingno=0.5.0=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.0.1=py36h3010b51_0
mkl_random=1.1.1=py36h0573a6f_0
multimethod=1.4=pypi_0
munch=2.5.0=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
networkx=2.5.1=pypi_0
notebook=6.4.3=pypi_0
numba=0.53.1=pypi_0
numexpr=2.6.5=py36h7bf3b9c_0
numpy=1.19.5=pypi_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pypi_0
pandas=1.1.5=pypi_0
pandas-profiling=3.0.0=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.0=py_0
pexpect=4.8.0=py36_0
phik=0.12.0=pypi_0
pickleshare=0.7.5=py36_0
pillow=8.3.1=pypi_0
pip=21.1.3=py36h06a4308_0
pretrainedmodels=0.7.4=pypi_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.8=py_0
psutil=5.8.0=pypi_0
ptyprocess=0.6.0=py36_0
pycparser=2.20=pypi_0
pydantic=1.8.2=pypi_0
pygments=2.7.1=py_0
pynndescent=0.5.4=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.1=py_0
pytorchcv=0.0.66=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=19.0.2=py36he6710b0_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=pypi_0
scikit-multilearn=0.2.0=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=52.0.0=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tangled-up-in-unicode=0.1.0=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.2.0=pyhb85f177_0
tifffile=2020.9.3=pypi_0
tk=8.6.10=hbc83047_0
torch=1.7.1=pypi_0
torchgeometry=0.1.2=pypi_0
torchvision=0.8.2=pypi_0
tornado=6.1=pypi_0
tqdm=4.61.2=pypi_0
traitlets=4.3.3=py36_0
typing-extensions=3.10.0.0=pypi_0
umap-learn=0.5.1=pypi_0
urllib3=1.26.6=pypi_0
visions=0.7.1=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.5.0=pypi_0
zlib=1.2.11=h7b6447c_3
